Move `TLAW::Params::Param#make` to `TLAW::Params#make`
`Param` shouldn’t have to know about subclasses
